+++ src/main/kotlin/net/ccbluex/liquidbounce/features/command/commands/client/CommandFriend.kt
@@ -23,7 +23,7 @@ import net.ccbluex.liquidbounce.features.command.CommandException
 import net.ccbluex.liquidbounce.features.command.CommandFactory
 import net.ccbluex.liquidbounce.features.command.builder.CommandBuilder
 import net.ccbluex.liquidbounce.features.command.builder.ParameterBuilder
-import net.ccbluex.liquidbounce.features.command.builder.Parameters
+import net.ccbluex.liquidbounce.features.command.builder.playerName
 import net.ccbluex.liquidbounce.features.misc.FriendManager
 import net.ccbluex.liquidbounce.utils.client.*
 import net.minecraft.text.ClickEvent
@@ -188,7 +188,7 @@ object CommandFriend : CommandFactory {
         return CommandBuilder
             .begin("add")
             .parameter(
-                Parameters.playerName()
+                ParameterBuilder.playerName()
                     .required()
                     .build()
             )
