+++ src/main/kotlin/net/ccbluex/liquidbounce/features/command/commands/client/CommandToggle.kt
@@ -21,7 +21,8 @@ package net.ccbluex.liquidbounce.features.command.commands.client
 import net.ccbluex.liquidbounce.features.command.Command
 import net.ccbluex.liquidbounce.features.command.CommandFactory
 import net.ccbluex.liquidbounce.features.command.builder.CommandBuilder
-import net.ccbluex.liquidbounce.features.command.builder.Parameters
+import net.ccbluex.liquidbounce.features.command.builder.ParameterBuilder
+import net.ccbluex.liquidbounce.features.command.builder.module
 import net.ccbluex.liquidbounce.features.module.ClientModule
 import net.ccbluex.liquidbounce.utils.client.MessageMetadata
 import net.ccbluex.liquidbounce.utils.client.chat
@@ -40,7 +41,7 @@ object CommandToggle : CommandFactory {
             .begin("toggle")
             .alias("t")
             .parameter(
-                Parameters.module()
+                ParameterBuilder.module()
                     .required()
                     .build()
             )
