+++ src/main/kotlin/net/ccbluex/liquidbounce/features/command/commands/client/client/CommandClientConfigSubcommand.kt
@@ -24,7 +24,7 @@ import net.ccbluex.liquidbounce.config.gson.adapter.toUnderlinedString
 import net.ccbluex.liquidbounce.config.types.nesting.Configurable
 import net.ccbluex.liquidbounce.features.command.builder.CommandBuilder
 import net.ccbluex.liquidbounce.features.command.builder.ParameterBuilder
-import net.ccbluex.liquidbounce.features.command.builder.Parameters
+import net.ccbluex.liquidbounce.features.command.builder.rootConfigurables
 import net.ccbluex.liquidbounce.features.module.ModuleManager.modulesConfigurable
 import net.ccbluex.liquidbounce.features.module.modules.render.ModuleHud
 import net.ccbluex.liquidbounce.utils.client.*
@@ -53,7 +53,7 @@ object CommandClientConfigSubcommand {
 
     private fun backupSubcommand() = CommandBuilder.begin("backup")
         .parameter(
-            Parameters.rootConfigurables()
+            ParameterBuilder.rootConfigurables()
                 .optional()
                 .build()
         )
@@ -114,7 +114,7 @@ object CommandClientConfigSubcommand {
     private fun resetSubCommand() = CommandBuilder
         .begin("reset")
         .parameter(
-            Parameters.rootConfigurables()
+            ParameterBuilder.rootConfigurables()
                 .optional()
                 .build()
         )
