+++ src/main/kotlin/net/ccbluex/liquidbounce/features/command/commands/ingame/creative/CommandItemEnchant.kt
@@ -23,7 +23,7 @@ import net.ccbluex.liquidbounce.features.command.CommandException
 import net.ccbluex.liquidbounce.features.command.CommandFactory
 import net.ccbluex.liquidbounce.features.command.builder.CommandBuilder
 import net.ccbluex.liquidbounce.features.command.builder.ParameterBuilder
-import net.ccbluex.liquidbounce.features.command.builder.Parameters
+import net.ccbluex.liquidbounce.features.command.builder.enchantment
 import net.ccbluex.liquidbounce.features.module.MinecraftShortcuts
 import net.ccbluex.liquidbounce.utils.client.MessageMetadata
 import net.ccbluex.liquidbounce.utils.client.chat
@@ -65,7 +65,7 @@ object CommandItemEnchant : CommandFactory, MinecraftShortcuts {
             .subcommand(
                 CommandBuilder
                     .begin("add")
-                    .parameter(Parameters.enchantment().required().build())
+                    .parameter(ParameterBuilder.enchantment().required().build())
                     .parameter(levelParameter.build())
                     .handler { command, args ->
                         val enchantmentName = args[0] as String
@@ -88,7 +88,7 @@ object CommandItemEnchant : CommandFactory, MinecraftShortcuts {
             .subcommand(
                 CommandBuilder
                     .begin("remove")
-                    .parameter(Parameters.enchantment().required().build())
+                    .parameter(ParameterBuilder.enchantment().required().build())
                     .handler { command, args ->
                         val enchantmentName = args[0] as String
