+++ src/main/kotlin/net/ccbluex/liquidbounce/features/command/commands/ingame/creative/CommandItemGive.kt
@@ -23,7 +23,7 @@ import net.ccbluex.liquidbounce.features.command.CommandException
 import net.ccbluex.liquidbounce.features.command.CommandFactory
 import net.ccbluex.liquidbounce.features.command.builder.CommandBuilder
 import net.ccbluex.liquidbounce.features.command.builder.ParameterBuilder
-import net.ccbluex.liquidbounce.features.command.builder.Parameters
+import net.ccbluex.liquidbounce.features.command.builder.item
 import net.ccbluex.liquidbounce.utils.client.*
 import net.ccbluex.liquidbounce.utils.item.createItem
 import net.minecraft.network.packet.c2s.play.CreativeInventoryActionC2SPacket
@@ -39,7 +39,7 @@ object CommandItemGive : CommandFactory {
         return CommandBuilder
             .begin("give")
             .requiresIngame()
-            .parameter(Parameters.item().required().build())
+            .parameter(ParameterBuilder.item().required().build())
             .parameter(
                 ParameterBuilder
                     .begin<Int>("amount")
