+++ src/main/kotlin/net/ccbluex/liquidbounce/features/command/commands/module/CommandInvsee.kt
@@ -23,7 +23,8 @@ import net.ccbluex.liquidbounce.features.command.Command
 import net.ccbluex.liquidbounce.features.command.CommandException
 import net.ccbluex.liquidbounce.features.command.CommandFactory
 import net.ccbluex.liquidbounce.features.command.builder.CommandBuilder
-import net.ccbluex.liquidbounce.features.command.builder.Parameters
+import net.ccbluex.liquidbounce.features.command.builder.ParameterBuilder
+import net.ccbluex.liquidbounce.features.command.builder.playerName
 import net.ccbluex.liquidbounce.features.module.modules.misc.ModuleInventoryTracker
 import net.ccbluex.liquidbounce.utils.client.mc
 import net.ccbluex.liquidbounce.utils.client.network
@@ -47,7 +48,7 @@ object CommandInvsee : CommandFactory {
             .begin("invsee")
             .requiresIngame()
             .parameter(
-                Parameters.playerName()
+                ParameterBuilder.playerName()
                     .required()
                     .build()
             )
